+++ .github/CONTRIBUTING.md
@@ -1,10 +1,10 @@
-# Welcome to GenAI-PoD Contributing Guide
+# Welcome to genai Contributing Guide
 
-Thank you for taking the time to contribute to the [GenAI-PoD](https://github.com/LeonhardSchwertfeger/GenAI-PoD/) project!
+Thank you for taking the time to contribute to the [genai](https://github.com/LeonhardSchwertfeger/genai/) project!
 
 Please make sure to read our [Code of Conduct](./CODE_OF_CONDUCT.md) to ensure a respectful and welcoming community.
 
-If you have questions, suggestions, or topics to discuss, feel free to participate in [GenAI-PoD/discussions](https://github.com/LeonhardSchwertfeger/GenAI-PoD/discussions). We look forward to engaging discussions, exchanging strategies, and fostering knowledge sharing!
+If you have questions, suggestions, or topics to discuss, feel free to participate in [genai/discussions](https://github.com/LeonhardSchwertfeger/genai/discussions). We look forward to engaging discussions, exchanging strategies, and fostering knowledge sharing!
 
 ## Getting Started 🤖
 
@@ -14,11 +14,11 @@ This guide provides an overview of the contribution workflow, including opening 
 
 #### Create a New Issue
 
-If you encounter an issue not listed in the troubleshooting section of [README.md](https://github.com/LeonhardSchwertfeger/GenAI-PoD#readme) or [documentation](), feel free to create a new [issue](https://github.com/LeonhardSchwertfeger/GenAI-PoD/issues). If the issue or feature request does not already exist, this is the best place to share it.
+If you encounter an issue not listed in the troubleshooting section of [README.md](https://github.com/LeonhardSchwertfeger/genai#readme) or [documentation](), feel free to create a new [issue](https://github.com/LeonhardSchwertfeger/genai/issues). If the issue or feature request does not already exist, this is the best place to share it.
 
 #### Solve an Issue
 
-Browse through the [existing issues](https://github.com/LeonhardSchwertfeger/GenAI-PoD/issues) to find one that interests you. You can also filter issues using labels to focus on specific areas.
+Browse through the [existing issues](https://github.com/LeonhardSchwertfeger/genai/issues) to find one that interests you. You can also filter issues using labels to focus on specific areas.
 
 ### Make Changes
 
@@ -29,15 +29,15 @@ Browse through the [existing issues](https://github.com/LeonhardSchwertfeger/Gen
    Clone your forked repository locally:
 
    ```bash
-   git clone https://github.com/<your-username>/GenAI-PoD.git
-   cd GenAI-PoD
+   git clone https://github.com/<your-username>/genai.git
+   cd genai
    ```
 
 3. **Set Up the Upstream Remote**
    Add the original repository as the upstream remote:
 
    ```bash
-   git remote add upstream https://github.com/LeonhardSchwertfeger/GenAI-PoD.git
+   git remote add upstream https://github.com/LeonhardSchwertfeger/genai.git
    ```
 
 4. **Synchronize Your Fork**
@@ -107,4 +107,4 @@ Open a pull request (PR) from your fork to the original repository. Make sure to
 
 ### Your PR is Merged! 🏅
 
-Congratulations! Thank you for contributing to GenAI-PoD. Feel free to pick up the next issue or suggest additional improvements. 🚀
+Congratulations! Thank you for contributing to genai. Feel free to pick up the next issue or suggest additional improvements. 🚀
