+++ .github/pull_request_template.md
@@ -8,9 +8,9 @@ What has being changed - provide code snippets, examples, screenshots, lists -
 any kind of information hat helps to understand the update.
 
 **Make sure to address all topics of the [self-review
-checklist](https://github.com/LeonhardSchwertfeger/GenAI-PoD/blob/master/.github/self-review.md).**
+checklist](https://github.com/LeonhardSchwertfeger/genai/blob/master/.github/self-review.md).**
 
 Closes: Don't forget to link an existing issue for your change. If there is no
 open issue for your change - create one to make it more likely that his update
 will be accepted:
-[GenAI-PoD/issues/new/choose](https://github.com/LeonhardSchwertfeger/GenAI-PoD/issues/new/choose).
+[genai/issues/new/choose](https://github.com/LeonhardSchwertfeger/genai/issues/new/choose).
