+++ .pre-commit-config.yaml
@@ -96,7 +96,7 @@ repos:
           - --config-file=pyproject.toml
           - --install-types
           - --non-interactive
-          - genai_pod
+          - genai
   - repo: https://github.com/psf/black
     rev: 24.10.0
     hooks:
