+++ genai/resources/scaling_adjustments.json
