+++ genai/resources/vexel_forbidden_words.json
