+++ genai/uploaders/redbubble.py
@@ -42,7 +42,7 @@ from selenium.webdriver.common.keys import Keys
 from seleniumbase import SB  # type: ignore[import]
 from tqdm import tqdm
 
-from genai_pod.utils import chromedata
+from genai.utils import chromedata
 
 logging.basicConfig(level=logging.INFO)
